+++ lib/helpers.py
@@ -1,11 +1,11 @@
 """
 /lib/helpers.py
-Server Monitoring System v8.7.0
+Server Monitoring System v8.8.0
 Copyright (c) 2025 Aleksandr Sukhanov
 License: MIT
 Auxiliary utilities
 Система мониторинга серверов
-Версия: 8.7.0
+Версия: 8.8.0
 Автор: Александр Суханов (c)
 Лицензия: MIT
 Вспомогательные утилиты