+++ main.py
@@ -1,12 +1,12 @@
 #!/usr/bin/env python3
 """
 /main.py
-Server Monitoring System v8.7.0
+Server Monitoring System v8.8.0
 Copyright (c) 2025 Aleksandr Sukhanov
 License: MIT
 Main launch module
 Система мониторинга серверов
-Версия: 8.7.0
+Версия: 8.8.0
 Автор: Александр Суханов (c)
 Лицензия: MIT
 Основной модуль запуска