+++ modules/resources.py
@@ -1,11 +1,11 @@
 """
 /app/modules/resources.py
-Server Monitoring System v8.7.0
+Server Monitoring System v8.8.0
 Copyright (c) 2025 Aleksandr Sukhanov
 License: MIT
 Server resource checking module
 Система мониторинга серверов
-Версия: 8.7.0
+Версия: 8.8.0
 Автор: Александр Суханов (c)
 Лицензия: MIT
 Модуль проверки ресурсов серверов